requirements.txt
Flask
line-bot-sdk
Pillow
pyocr